requirements.txt
aiodns==3.1.1
aiofiles==23.1.0
aiogram==2.25.2
aiogram-utils==0.1.4
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
contourpy==1.1.0
crispy-bootstrap4==2022.1
cycler==0.11.0
DateTime==5.2
Django==4.2.5
django-crispy-forms==2.0
dnspython==2.4.2
et-xmlfile==1.1.0
filelock==3.12.4
Flask==2.3.2
fonttools==4.40.0
frozenlist==1.4.0
gdown==4.7.1
ggl==1.1.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gspread==5.12.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
magic-filter==1.0.12
MarkupSafe==2.1.3
matplotlib==3.7.1
mongoengine==0.24.2
multidict==6.0.4
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
Pillow==9.5.0
protobuf==4.25.0
psycopg2-binary==2.9.8
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.4.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.6.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
Telethon==1.32.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==2.3.6
yarl==1.9.2
zope.interface==6.1
